requirements.txt
telethon==1.17.5
python-dotenv==0.15.0
requests==2.24.0
psycopg2-binary==2.8.6
uuid==1.30